add wallManager

+++ Assets/Scripts/Level/Level.cs
@@ -2,17 +2,17 @@
 {
     private int _levelNumber;
     private int _countWall;
-    private int _wallSpeed;
+    private float _wallSpeed;
 
     public int LevelNumber { get => _levelNumber; private set => _levelNumber = value; }
     public int CountWall { get => _countWall; private set => _countWall = value; }
-    public int WallSpeed { get => _wallSpeed; set => _wallSpeed = value; }
+    public float WallSpeed { get => _wallSpeed; set => _wallSpeed = value; }
 
     public Level()
     {
         _levelNumber = 0;
         _countWall = 4;
-        _wallSpeed = 100;
+        _wallSpeed = 100f;
     }
 
     public void NextLevel()
@@ -21,7 +21,7 @@
         if (_levelNumber % 2 == 0)
         {
             _countWall++;
-            _wallSpeed += 5;
+            _wallSpeed += 5f;
         }
     }
 }
